added simple user operation/rigth control

+++ fstore/fssrv/fsrec/bstore.go
@@ -5,26 +5,48 @@
 package fsrec
 
 import (
+    "errors"
     "ndstore/dscom"
 )
 
+const BStateNormal      string = "normal"
+const BStateDisabled    string = "disabled"
+const BStateWrong       string = "wrong"
+
 func (store *Store) SeedBStores() error {
     var err error
+
+    bStores, err := store.reg.ListBStoreDescrs()
+    if err != nil {
+        return err
+    }
+
+    if len(bStores) > 0 {
+        return err
+    }
+
     const address   = "127.0.0.1"
-    var port      = "5100"
     const login     = "admin"
     const pass      = "admin"
-    _, err = store.reg.AddBStoreDescr(address, port, login, pass, StateEnabled)
-    port      = "5101"
-    _, err = store.reg.AddBStoreDescr(address, port, login, pass, StateEnabled)
-    port      = "5102"
-    _, err = store.reg.AddBStoreDescr(address, port, login, pass, StateEnabled)
+    ports := []string{ "5101", "5102", "5103" }
+    for _, port := range ports {
+        _, err = store.reg.AddBStoreDescr(address, port, login, pass, BStateNormal)
+        if err != nil {
+            return err
+        }
+    }
     return err
 }
 
-func (store *Store) AddBStore(address, port, login, pass string) error {
+func (store *Store) AddBStore(userName, address, port, login, pass string) error {
     var err error
-    _, err = store.reg.AddBStoreDescr(address, port, login, pass, StateEnabled)
+
+    role, err := store.reg.GetUserRole(userName)
+    if role != URoleAdmin {
+        return errors.New("insufficient rights for adding bStore")
+    }
+
+    _, err = store.reg.AddBStoreDescr(address, port, login, pass, BStateNormal)
     if err != nil {
         return err
     }
@@ -37,27 +59,47 @@ func (store *Store) GetBStore(address, port string) (*dscom.BStoreDescr, error) 
     return bStore, err
 }
 
-func (store *Store) UpdateBStore(address, port, login, pass string) error {
+func (store *Store) UpdateBStore(userName, address, port, login, pass string) error {
     var err error
-    ok, err := checkPass(pass)
+
+    role, err := store.reg.GetUserRole(userName)
+    if role != URoleAdmin {
+        return errors.New("insufficient rights for updating bStore")
+    }
+
+    ok, err := validatePass(pass)
     if !ok {
         return err
     }
-    err = store.reg.UpdateBStoreDescr(address, port, login, pass, StateEnabled)
+    err = store.reg.UpdateBStoreDescr(address, port, login, pass, BStateNormal)
     return err
 }
 
-func (store *Store) ListBStores() ([]*dscom.BStoreDescr, error) {
+func (store *Store) ListBStores(userName string) ([]*dscom.BStoreDescr, error) {
     var err error
-    bStores, err := store.reg.ListBStoreDescrs()
+    bStores := make([]*dscom.BStoreDescr, 0)
+
+    role, err := store.reg.GetUserRole(userName)
+    if role != URoleAdmin {
+        return bStores, errors.New("insufficient rights for listing bStores")
+    }
+
+    bStores, err = store.reg.ListBStoreDescrs()
     //for i := range BStores {
     //    BStores[i].Pass = "xxxxx"
     //}
+    if err != nil {
+        return bStores, err
+    }
     return bStores, err
 }
 
-func (store *Store) DeleteBStore(address, port string) error {
+func (store *Store) DeleteBStore(userName, address, port string) error {
     var err error
+    role, err := store.reg.GetUserRole(userName)
+    if role != URoleAdmin {
+        return errors.New("insufficient rights for delete bStore")
+    }
     err = store.reg.DeleteBStoreDescr(address, port)
     return err
 }
